requirements.txt
pip>=25.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pikepdf==8.15.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.23.8
numpy==1.24.3
opencv-python-headless==4.8.1.78